Escaping select special characters.
At this point it's not clear to me what is the problem with the LIMS API and UTF-8 chars. It seemed like it was only in researcher names on the Process resource, but now I'm getting errors on UDF values again. It's probably to do with the encoding on the client.